Remove API_HOST variable from swagger
The deployment scripts were allowing the swagger `Execute` requests to be set to customizable "API host". Since the scala API is what provides the swagger docs there is no need to override the backend API host.

This patch purges the API_HOST variable and the swagger docs will now, by default, send the request to the host that where it's running.

If a different host is needed in the future, use this approach https://github.com/iheartradio/play-swagger#how-do-i-use-a-different-host-for-different-environment