Add 'Use Host Network' Option to Deployment Script
Added a 'use host network' option (--useHostNetwork) to the deployment script.
This option allows Docker to use the host network for the mr-postgis container
instead of the default bridged network.
When this option is used, the script will apply the '--network=host' Docker argument instead of '-p <ports> --network mrnet'.